Move createCase to src/utils
